requirements.txt
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.4.0
tiktoken==0.3.3
tqdm==4.65.0
